refactor(web): phase 1 - type safety, TOC bug, lazy reply query
- Fix Promise<any> in blog.ts and comment.ts server functions with proper serializable types
- Move CommentWithRelations from types/misc.ts to types/comment.ts
- Fix any cast in comment-replies.tsx (use CommentWithRelations)
- Fix TOC hash double-# bug (normalize item.url before prepending #)
- Add enabled: isOpenReplies to comment-replies useQuery to avoid unnecessary fetches
- Fix ArticleAuthor prop type to accept null author
- Fix relatedArticles map type in article detail route
