+++ apps/web/src/components/blog/comment/comment-replies.tsx
@@ -4,6 +4,7 @@ import { Loader2Icon } from 'lucide-react';
 import { useCommentContext } from '@/contexts/comment';
 import { queryKeys } from '@/lib/query-keys';
 import { $getAllComments } from '@/lib/server';
+import type { CommentWithRelations } from '@/types/misc';
 import CommentItem from './comment-item';
 
 interface CommentRepliesProps {
@@ -15,6 +16,7 @@ export default function CommentReplies({ articleSlug }: Readonly<CommentRepliesP
 
   const { data: comments, isLoading } = useQuery({
     queryKey: queryKeys.comment.byArticleAndParent(comment.comment.articleId, comment.comment.id),
+    enabled: isOpenReplies,
     queryFn: () =>
       $getAllComments({
         data: {
@@ -27,8 +29,9 @@ export default function CommentReplies({ articleSlug }: Readonly<CommentRepliesP
   return (
     <div>
       {isOpenReplies && !isLoading ? (
-        // biome-ignore lint/suspicious/noExplicitAny: Drizzle relation types from $getAllComments
-        comments?.map((reply: any) => <CommentItem articleSlug={articleSlug} comment={reply} key={reply.comment.id} />)
+        comments?.map((reply: CommentWithRelations) => (
+          <CommentItem articleSlug={articleSlug} comment={reply} key={reply.comment.id} />
+        ))
       ) : (
         <Button className="px-0" onClick={() => setIsOpenReplies(true)} type="button" variant="link">
           Show all {comment.repliesCount} replies
