+++ apps/web/src/routes/(public)/blog/$articleId.tsx
@@ -316,12 +316,12 @@ function RouteComponent() {
           >
             <h2 className="mb-6 font-bold font-heading text-2xl tracking-tight">Related Posts</h2>
             <div className="grid gap-6 sm:grid-cols-2 md:grid-cols-3">
-              {article.relatedArticles.map((relatedArticle: Parameters<typeof ArticleCard>[0]['article']) => (
+              {article.relatedArticles.map((relatedArticle) => (
                 <ArticleCard
                   article={{
                     ...relatedArticle,
-                    viewCount: relatedArticle.viewCount ?? 0,
-                    likesCount: relatedArticle.likesCount ?? 0,
+                    viewCount: 0,
+                    likesCount: 0,
                   }}
                   key={relatedArticle.slug}
                 />
